♻️ refactor copy to clipboard
