+++ src/App.tsx
@@ -13,6 +13,7 @@ import Screen03Project from "./pages/Screen03Project";
 import { Outlet } from "react-router";
 import Screen04Education from "./pages/Screen04Education";
 import Footer from "./pages/Footer";
+import { Tooltip } from "react-tooltip";
 
 export const QueryContext = createContext<ContextEntry | null>(null);
 
@@ -61,6 +62,11 @@ export default function App() {
     });
   };
 
+  const copyToClipBoard = (id: string): void => {
+    const content: string = document.getElementById(id)?.textContent ?? "";
+    navigator.clipboard.writeText(content);
+  };
+
   useEffect(() => {
     window.addEventListener("scroll", getScrollPos);
 
@@ -68,7 +74,7 @@ export default function App() {
   }, []);
 
   return (
-    <QueryContext.Provider value={{ isDetails, setIsDetails }}>
+    <QueryContext.Provider value={{ isDetails, setIsDetails, copyToClipBoard }}>
       <div
         className="crelative flex min-h-screen w-full flex-col"
         onClick={() => closeTooltips()}
@@ -81,16 +87,22 @@ export default function App() {
         <Screen01AboutMe screenRef={screenRefs[0]} />
         <Screen02Career screenRef={screenRefs[1]} />
         <Screen03Project screenRef={screenRefs[2]} />
-        {/* <Screen04Education screenRef={screenRefs[3]} /> */}
+        <Screen04Education screenRef={screenRefs[3]} />
         <Footer />
       </div>
       <ContactIcons showContact={showContact} setShowContact={setShowContact} />
       <Outlet />
+      <Tooltip
+        id="icons-tooltip"
+        openOnClick={true}
+        closeEvents={{ mouseleave: true }}
+      />
     </QueryContext.Provider>
   );
 }
 
 export interface ContextEntry {
   isDetails: boolean;
+  copyToClipBoard: (id: string) => void;
   setIsDetails: React.Dispatch<React.SetStateAction<boolean>>;
 }
