+++ src/css/index.css
@@ -21,6 +21,7 @@
 
 body {
   background-color: #242424;
+  color: #f1f1f1;
 }
 
 .no-scroll {
